🔨 Refactor(#142): 잘못만든 상수파일  제거
