--- src/pages/constants.ts
@@ -1,16 +1,0 @@
-export const BUTTON_LIST: { id: number; label: string }[] = [
-  { id: 1, label: '조합형' },
-  { id: 2, label: '객관식' },
-  {
-    id: 3,
-    label: '단답형',
-  },
-  { id: 4, label: 'O, X' },
-];
-
-export const COKO_TEAM_INFO: { label: string; createBy: string[] }[] = [
-  { label: '만든 사람들', createBy: ['모던애자일 8기'] },
-  { label: '프론트', createBy: ['구도윤', '신현성'] },
-  { label: '백엔드', createBy: ['홍대경', '이건우', '전희나'] },
-  { label: '디자인', createBy: ['이채영'] },
-];
